Extracting md5 method to utils
